Collapse app/audio/ folder; split data into components/, systems into systems/ (refs #1200)
Mirrors the input/ collapse template (#1216) for the audio/ subfolder.

Moves:
| audio/audio_types.h     | components/audio.h        | SFX enum + SFXBank   |
| audio/music_context.h   | components/music.h        | MusicContext         |
| audio/audio_routing.h   | systems/audio_routing.h   | listeners + wiring   |
| audio/sfx_bank.h        | systems/sfx_bank.h        | bank init/unload API |
| audio/audio_dispatcher.cpp | systems/audio_dispatcher.cpp | wiring impl   |
| audio/sfx_bank.cpp      | systems/sfx_bank.cpp      | bank impl + SFXBank methods |

CMakeLists.txt: drops the AUDIO_SOURCES glob; the relocated .cpp files are
picked up by the existing SYSTEM_SOURCES glob. The app/audio/ directory is
deleted.

All call sites updated: app/game_loop.cpp, app/session/play_session.cpp,
app/components/audio_events.h, app/systems/{audio_system,haptic_system,
song_playback_system}.cpp, benchmarks/bench_systems.cpp, and the audio /
GPU-lifecycle / music / song-playback test files. Header comment paths and
docs (design-docs/architecture.md, beatmap-integration.md, audio-pipeline-spec.md)
also updated.

Validation: 901 Catch2 tests / 913 ctest cases pass with zero warnings under
both unity and non-unity builds (Clang -Wall -Wextra -Werror) on macOS.
