+++ app/game_loop.cpp
@@ -4,12 +4,12 @@
 #include "components/input.h"
 #include "components/game_state.h"
 #include "components/scoring.h"
-#include "audio/audio_types.h"
-#include "audio/sfx_bank.h"
-#include "audio/audio_routing.h"
+#include "components/audio.h"
+#include "systems/sfx_bank.h"
+#include "systems/audio_routing.h"
 #include "entities/settings.h"
 #include "components/rhythm.h"
-#include "audio/music_context.h"
+#include "components/music.h"
 #include "components/rendering.h"
 #include "components/test_player.h"
 #include "components/obstacle.h"
