+++ benchmarks/bench_systems.cpp
@@ -12,8 +12,8 @@
 #include "components/rendering.h"
 #include "components/particle.h"
 #include "components/rhythm.h"
-#include "audio/audio_types.h"
-#include "audio/audio_routing.h"
+#include "components/audio.h"
+#include "systems/audio_routing.h"
 #include "constants.h"
 #include "systems/all_systems.h"
 #include "systems/input_routing.h"
