+++ tests/test_helpers.h
@@ -10,10 +10,10 @@
 #include "components/scoring.h"
 #include "components/rendering.h"
 #include "components/particle.h"
-#include "audio/audio_types.h"
+#include "components/audio.h"
 #include "components/haptics.h"
 #include "components/audio_events.h"
-#include "audio/audio_routing.h"
+#include "systems/audio_routing.h"
 #include "entities/settings.h"
 #include "components/rhythm.h"
 #include "util/rhythm_math.h"
